MT#55283 fix int size issues
(cherry picked from commit 675eb7bea6f37d72cadb6d57f5e3011485e3af7a)

+++ lib/bufferpool.h
@@ -6,7 +6,7 @@
 #define BUFFERPOOL_ALIGNMENT (sizeof(void *)) // bytes
 #define BUFFERPOOL_ALIGN(x) (((x + BUFFERPOOL_ALIGNMENT - 1) / BUFFERPOOL_ALIGNMENT) * BUFFERPOOL_ALIGNMENT)
 
-#define BUFFERPOOL_SHARD_SIZE (1LL<<24) // 16 MB, must be a power of two
+#define BUFFERPOOL_SHARD_SIZE ((size_t) (1LL<<24)) // 16 MB, must be a power of two
 #define BUFFERPOOL_OVERHEAD BUFFERPOOL_ALIGN(sizeof(void *)) // storage space not available
 
 #define BUFFERPOOL_BOTTOM_MASK (BUFFERPOOL_SHARD_SIZE - 1)
